Inline AppUpdater into App::UpdateCommand
